src/main/java/com/example/copsboot/security/oauth2/ApplicationUserDetailsService.java
package com.example.copsboot.security.oauth2;

import com.example.copsboot.domain.user.User;
import com.example.copsboot.repository.UserRepository;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;

public class ApplicationUserDetailsService implements UserDetailsService {
    private final UserRepository userRepository;

    public ApplicationUserDetailsService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userRepository.findByEmailIgnoreCase(username)
                .orElseThrow(() -> new UsernameNotFoundException(String.format("User with email %s could not be found",username)));

        return new ApplicationUserDetails(user);
    }
}
